+++ tests/src/glr/regressions/mod.rs
@@ -1,1 +1,2 @@
 mod issue_16_subtract_overflow_panic;
+mod issue_22_panic_get_conflicts;